[SP] update NMEAparser code. change variable float to double. change taskcreate stack size.

+++ PIC-MPLab/PIC24-Proj/ePaper-trial/src/services/NMEAparser/NMEAcalculation.h
@@ -17,8 +17,8 @@
 
 
 
-typedef    	float		nmeaCalcRET;
-///typedef     double		nmeaCalcRET;
+///typedef    	float		nmeaCalcRET;
+typedef     double		nmeaCalcRET;
 
 
 nmeaCalcRET nmeaCalcDistanceIn2Point( NMEA_GPS *pPoint1, NMEA_GPS *pPoint2 );
